backend/requirements.txt
asyncio==3.4.3
boto3==1.36.2
botocore==1.36.2
certifi==2024.8.30
fastapi==0.115.6
fastapi-health==0.4.0
google-api-core==2.24.0
google-auth==2.37.0
google_auth_oauthlib==1.2.1
google-cloud-core==2.4.1
json-repair==0.30.2
pip-install==1.3.5
langchain==0.3.15
langchain-aws==0.2.11
langchain-anthropic==0.3.3
langchain-fireworks==0.2.6
langchain-community==0.3.15
langchain-core==0.3.31
langchain-experimental==0.3.4
langchain-google-vertexai==2.0.11
langchain-groq==0.2.3
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langchain-huggingface==0.1.2
langdetect==1.0.9
langsmith==0.2.11
langserve==0.3.1
neo4j-rust-ext
nltk==3.9.1
openai==1.59.9
opencv-python==4.10.0.84
psutil==6.1.0
pydantic==2.9.2
python-dotenv==1.0.1
python-magic==0.4.27
PyPDF2==3.0.1
PyMuPDF==1.24.14
starlette==0.41.3
sse-starlette==2.1.3
starlette-session==0.4.3
tqdm==4.67.1
unstructured[all-docs]
unstructured==0.16.11
unstructured-client==0.28.1
unstructured-inference==0.8.1
urllib3==2.2.2
uvicorn==0.32.1
gunicorn==23.0.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
youtube-transcript-api==0.6.3
zipp==3.17.0
sentence-transformers==3.3.1
google-cloud-logging==3.11.3
pypandoc==1.13
graphdatascience==1.12
Secweb==1.11.0
ragas==0.2.11
rouge_score==0.1.2
langchain-neo4j==0.3.0
pypandoc-binary==1.15
chardet==5.2.0
sentence-transformers==3.3.1